Treat random rotation variance as a percentage instead of degrees.
Previously, the "fixed random rotation" feature was based on some complicated logic to ensure that the variance could be expressed in degrees. Besides leading to discrepancies in saving/loading, it wasn't true to the game's behavior. This change reworks the logic (and UI) to express the variance in a percentage of the base rotation angle. This change fixes #5.